requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
build==1.2.1
    # via pip-tools
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via -r requirements.in
distlib==0.3.8
    # via virtualenv
filelock==3.15.4
    # via virtualenv
flake8==7.1.1
    # via -r requirements.in
identify==2.6.0
    # via pre-commit
iniconfig==2.0.0
    # via pytest
jsonschema==4.23.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
markupsafe==2.1.5
    # via werkzeug
marshmallow==3.22.0
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mysql-connector-python==9.0.0
    # via -r requirements.in
nodeenv==1.9.1
    # via pre-commit
packaging==24.1
    # via
    #   build
    #   marshmallow
    #   pytest
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.8.0
    # via -r requirements.in
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==8.3.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via pre-commit
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
toolz==0.12.1
    # via -r requirements.in
virtualenv==20.26.3
    # via pre-commit
werkzeug==3.0.4
    # via -r requirements.in
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
